+++ tsdb/seriesfile/metrics.go
@@ -1,4 +1,4 @@
-package tsdb
+package seriesfile
 
 import (
 	"sort"