+++ ui/src/timeMachine/reducers/index.ts
@@ -803,6 +803,7 @@ export const timeMachineReducer = (
 
         tag.key = key
         tag.values = []
+        tag.aggregateFunctionType = 'filter'
       })
     }
 
@@ -812,7 +813,6 @@ export const timeMachineReducer = (
         const draftQuery = draftState.draftQueries[draftState.activeQueryIndex]
 
         draftQuery.builderConfig.tags[index].values = values
-
         buildActiveQuery(draftState)
       })
     }